requirements.txt
amqp==2.1.4
appdirs==1.4.2
billiard==3.5.0.2
bootstrap3-datetime==2.4
celery==4.0.2
Django==1.10.5
django-allauth==0.31.0
django-celery-results==1.0.1
django-ckeditor==5.2.2
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-ipware==1.1.6
django-widget-tweaks==1.4.1
kombu==4.0.2
oauthlib==2.0.2
packaging==16.8
pyparsing==2.1.10
python-openid==2.2.5
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
vine==1.1.3
